修复菜式组成成分转型问题。

+++ src/main/java/com/bkit/fatdown/mappers/TbFoodElementRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbFoodElementRelationMapper {
-    Integer countByExample(TbFoodElementRelationExample example);
+    long countByExample(TbFoodElementRelationExample example);
 
     int deleteByExample(TbFoodElementRelationExample example);
